Refactor: "code refactoring"
변경점
1. recent와 trend는 post와 연관된 내용이므로 post 내부로 합침
