+++ src/main/kotlin/com/example/velog/domain/post/controller/RecentController.kt
@@ -1,7 +1,7 @@
-package com.example.velog.domain.trend.controller
+package com.example.velog.domain.post.controller
 
 import com.example.velog.domain.post.dto.PostResponseDto
-import com.example.velog.domain.trend.service.TrendService
+import com.example.velog.domain.post.service.PostService
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.HttpStatus
@@ -11,11 +11,11 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 
-@Tag(name = "trend", description = "조회수가 높은 게시글을 보여주는 API")
-@RequestMapping("/trend")
+@Tag(name = "recent", description = "최근에 만들어진 게시글을 보여주는 API")
+@RequestMapping("/recent")
 @RestController
-class TrendController(
-    private val trendService: TrendService
+class RecentController(
+    private val postService: PostService
 ) {
     @Operation(summary = "게시글 목록 조회", description = "게시글 목록을 가져옵니다.")
     @GetMapping
@@ -28,6 +28,6 @@ class TrendController(
     ): ResponseEntity<List<PostResponseDto>> {
         return ResponseEntity
             .status(HttpStatus.OK)
-            .body(trendService.getTrendList(page, size))
+            .body(postService.getRecentList(page, size))
     }
 }
