+++ src/main/kotlin/com/example/velog/domain/post/repository/PostRepository.kt
@@ -5,4 +5,6 @@ import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PostRepository : JpaRepository<PostEntity, Long> {
+    fun findAllByOrderByCreateAtDesc(pageable: Pageable): List<PostEntity>
+    fun findAllByOrderByViewsDesc(pageable: Pageable): List<PostEntity>
 }
