--- src/main/kotlin/com/example/velog/domain/recent/service/RecentService.kt
@@ -1,8 +1,0 @@
-package com.example.velog.domain.recent.service
-
-import com.example.velog.domain.post.dto.PostResponseDto
-
-interface RecentService {
-
-    fun getRecentList(page: Int, size: Int): List<PostResponseDto>
-}
