--- src/main/kotlin/com/example/velog/domain/trend/service/TrendService.kt
@@ -1,7 +1,0 @@
-package com.example.velog.domain.trend.service
-
-import com.example.velog.domain.post.dto.PostResponseDto
-
-interface TrendService {
-    fun getTrendList(page: Int, size: Int): List<PostResponseDto>
-}
